+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.0.10] - 2021-12-21
+### Added
+- flag to parse annotations using strings.HasSuffix to create button: -S or --annotations-suffix-link
+- flag to exclude annotations when creating buttons: -E or --annotations-suffix-exclude
+
 ## [0.0.9] - 2020-11-30
 ### Added
 - Annotations and labels parsing checking